+++ Projektikansio/SharedLib/matkakohdeDTO.cs
@@ -2,7 +2,7 @@
 {
     public class matkakohdeDTO
     {
-        public int idmatkakohde { get; set; }
+        public long idmatkakohde { get; set; }
         public string kohdenimi { get; set; }
         public string maa { get; set; }
         public string paikkakunta { get; set; }
